+++ counter_slider/CHANGELOG.md
@@ -3,7 +3,7 @@
  - More dynamic demo, where you can preview the component.
  - Fixed documentation. #TODO
  - Renamed `setValue` to `onChanged`.
- - Renamed `buttonBorderGap` to `gapSize` #TODO
+ - Renamed `buttonBorderGap` to `gapSize`
  - Change fixed width + height to optional, default unconstrained. #TODO
  - Fixed `borderSize: 0` still visible. #TODO